saga_campaign_weaver/requirements.txt
fastapi
uvicorn
pydantic
langchain
langchain-openai
httpx
python-dotenv
